Update generateTiffThumbnails.js
Fix hardcoded path (whoops!)
TIFF thumbnails are likely to be removed anyways. We want precomputed files for ALL 3D images.

This means we need some sort of image pipeline - ex, all files to NIFTI, NIFTI to thumbnails, NIFTI to precomputed.

There might be a way to use precomputed to generate the thumbnails as well. Perhaps a library for using precomputeds? That would prevent the need to load the entire NIFTI file twice.